Partially revert "[mms] Fix parsing icalendar attributes that contain duplicate keys."
This reverts commit c4f07c6ca953f5d04603b9e10ad629e701ac103e.

getAttribute()'s API is too loose. There *really* should be a way to
force either a string value (lib makes best choice decision as to which
value, if multiple, should be returned) or array.  But this is not
possible - it is up to caller to explicitly check for return type :(

Compromise: new getAttributeSingle() method